Revert "Updated gui, added artifact to create jar"
This reverts commit 563a7e5fa590570236c9b6549f3522de926087ed.

+++ src/main/java/com/blag/harmonysearch/gui/HarmonySearcherGui.java
@@ -32,7 +32,7 @@ public class HarmonySearcherGui extends HarmonySearcher
         super(function, harmonyMemorySize, maxImprovisationCount, harmonyMemoryConsiderationRatio, pitchAdjustmentRatio, argumentGenerationLimits);
         bestSolutions = FXCollections.observableArrayList();
         plot = new Plot();
-        plot.setParameters(function, argumentGenerationLimits);
+        plot.setParameters(function,argumentGenerationLimits);
     }
 
     /**
@@ -59,7 +59,7 @@ public class HarmonySearcherGui extends HarmonySearcher
                 if (harmonyMemory.count() > 0 && !harmonyMemory.getBestSolution().equals(currentBestSolution))
                 {
                     currentBestSolution = harmonyMemory.getBestSolution();
-                    bestSolutions.add(0, new SolutionGui((float) currentBestSolution.getValue(), currentBestSolution.getArguments(), improvisationCount));
+                    bestSolutions.add(0, new SolutionGui(currentBestSolution.getValue(), currentBestSolution.getArguments(), improvisationCount));
                     plot.drawPoint(currentBestSolution);
                 }
