+++ src/main/java/com/blag/harmonysearch/gui/SolutionGui.java
@@ -1,7 +1,6 @@
 package com.blag.harmonysearch.gui;
 
 import javafx.beans.property.SimpleDoubleProperty;
-import javafx.beans.property.SimpleFloatProperty;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
@@ -11,13 +10,13 @@ import lombok.Setter;
 @Setter
 public class SolutionGui
 {
-    private float value;
+    private double value;
     private double[] arguments;
     private int iterationNumber;
 
-    public SimpleFloatProperty getArgument(int whichOne)
+    public SimpleDoubleProperty getArgument(int whichOne)
     {
-        return new SimpleFloatProperty((float) arguments[whichOne]);
+        return new SimpleDoubleProperty(arguments[whichOne]);
     }
 
     @Override
@@ -25,7 +24,7 @@ public class SolutionGui
     {
         StringBuilder builder = new StringBuilder();
         builder.append("f(x) = ")
-                .append(String.valueOf((float) value))
+                .append(String.valueOf(value))
                 .append("\n");
 
 
@@ -38,7 +37,7 @@ public class SolutionGui
                 builder.append(String.valueOf(i + 1));
             }
             builder.append(" = ")
-                    .append(String.valueOf((float) arguments[i]))
+                    .append(String.valueOf(arguments[i]))
                     .append("\n");
         }
